requirements.txt
asgiref==3.5.0
certifi==2021.10.8
cloudinary==1.29.0
coverage==6.3.2
dj-database-url==0.5.0
Django==4.0.3
django-environ==0.8.1
django-widget-tweaks==1.4.12
gunicorn==20.1.0
Pillow==9.1.0
psycopg2==2.9.3
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
whitenoise==6.0.0
